% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoData}
\alias{demoData}
\title{A demo dataset for GeoMx Cancer Transcriptome Atlas (CTA) panel}
\format{
A NanoStringGeoMxSet S4 object with 8707 features and 88 samples
}
\usage{
data(demoData)
}
\description{
A demo dataset contains 88 ROIs and 8707 features
}
\examples{
data(demoData)

}
\keyword{datasets}
