% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidney}
\alias{kidney}
\title{A demo dataset for GeoMx Human Whole Transcriptome Atlas (WTA) panel}
\format{
A NanoStringGeoMxSet S4 object with 18642 features and 276 samples
}
\usage{
data(kidney)
}
\description{
A demo dataset contains 276 ROIs and 18642 features
}
\examples{
data(kidney)

}
\keyword{datasets}
