% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBthDEmod2}
\alias{NBthDEmod2}
\title{A demo example output list returned by function fitNBthDE}
\format{
A list
}
\usage{
data(NBthDEmod2)
}
\description{
A list used to demonstrate the function coefNBth
}
\examples{
data(NBthDEmod2)

}
\keyword{datasets}
