% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBthmDEmod2}
\alias{NBthmDEmod2}
\title{A demo example output list returned by function fitNBthmDE}
\format{
A list
}
\usage{
data(NBthmDEmod2)
}
\description{
A list used to demonstrate the function coefNBth
}
\examples{
data(NBthmDEmod2)

}
\keyword{datasets}
