% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{A demo dataset for GeoMx Human Whole Transcriptome Atlas (WTA) panel}
\format{
A set of lists used to test the function fitNBthDE
}
\usage{
data(test_data)
}
\description{
A demo dataset contains 276 ROIs and 18642 features
}
\examples{
data(test_data)

}
\keyword{datasets}
